refactor: Remove alumno id field, alter alumno dni and legajo fields
This commit removes the "id" field from the Alumno model and alters the "dni" and "legajo" fields. The "id" field is no longer needed as it is automatically generated by Django. The "dni" field is now set as the primary key, and the "legajo" field is set as unique. These changes improve the data structure and integrity of the Alumno model.