Fix issue with fetching and signing commitments
The code changes in `Compromiso.ts` and `Pagos.ts` fix an issue with fetching and signing commitments. The previous code was not correctly retrieving the user's DNI from the cookies, resulting in incorrect data being fetched and signed. The updated code now uses a hardcoded DNI value for testing purposes. Additionally, the `Pagos.ts` file now includes a new parameter `nro_transferencia` for sending the transfer number along with the payment information.

This commit also includes commented out code in `TablaPagos.tsx` that was causing a page reload. The code was commented out as it was not necessary and caused unnecessary page refreshes.